Implemented transfer function coeffs transferring to Filter class

+++ filter.h
@@ -39,8 +39,8 @@ private:
 protected:
     QVector< std::complex<float> > Poles;
     QVector< std::complex<float> > Zeros;
-    QVector<float> vec_B; // Transfer function numenator
-    QVector<float> vec_A; // and denominator
+    QVector<long double> vec_B; // Transfer function numenator
+    QVector<long double> vec_A; // and denominator
     QVector<RC_elements> Sections;
 
     Filter::FType ftype;
@@ -83,7 +83,7 @@ public:
 
     virtual bool calcFilter();
 
-    void set_TrFunc(QVector<float> a, QVector<float> b);
+    void set_TrFunc(QVector<long double> a, QVector<long double> b);
 
 };
